convert to External Module
