+++ libs/androidlib/databinding-api/src/mill/androidlib/databinding/AndroidDataBindingWorker.scala
@@ -1,6 +1,6 @@
 package mill.androidlib.databinding
 
-trait AndroidDataBinding {
+trait AndroidDataBindingWorker {
   def processResources(args: ProcessResourcesArgs): Unit
   def generateBindingSources(args: GenerateBindingSourcesArgs): Unit
 }
