+++ libs/androidlib/src/mill/androidlib/AndroidKotlinModule.scala
@@ -1,9 +1,17 @@
 package mill.androidlib
 
-import mill.api.{PathRef, Result}
+import mill.*
+import mill.api.{ModuleRef, PathRef, Result}
 import mill.javalib.{CoursierModule, Dep}
 import mill.kotlinlib.{Dep, DepSyntax, KotlinModule}
 import mill.{T, Task}
+import mill.androidlib.databinding.{
+  AndroidDataBindingWorker,
+  GenerateBindingSourcesArgs,
+  ProcessResourcesArgs,
+  AndroidDataBindingWorkerModule
+}
+import mill.util.Jvm
 
 // TODO expose Compose configuration options
 // https://kotlinlang.org/docs/compose-compiler-options.html possible options
@@ -13,6 +21,119 @@ trait AndroidKotlinModule extends KotlinModule with AndroidModule { outer =>
    * Enable Jetpack Compose support in the module. Default is `false`.
    */
   def androidEnableCompose: T[Boolean] = false
+
+  def enableViewBinding: Boolean = false
+
+  def enableDataBinding: Boolean = false
+
+  private def isBindingEnabled: Boolean = enableViewBinding || enableDataBinding
+
+  def androidDataBindingCompilerDeps: T[Seq[Dep]] = Task {
+    Seq(
+      mvn"androidx.databinding:databinding-compiler:8.13.0",
+      mvn"androidx.databinding:databinding-compiler-common:8.13.0"
+    )
+  }
+
+  def androidDataBindingWorkerModule: ModuleRef[AndroidDataBindingWorkerModule] =
+    ModuleRef(AndroidDataBindingWorkerModule)
+
+  def androidDataBindingCompilerClasspath: T[Seq[PathRef]] = Task {
+    defaultResolver().classpath(
+      Seq(
+        Dep.millProjectModule("mill-libs-androidlib-databinding")
+      ) ++ androidDataBindingCompilerDeps()
+    )
+  }
+
+  def androidDataBindingWorkerClassloader: Worker[ClassLoader] = Task.Worker {
+    Jvm.createClassLoader(
+      classPath = androidDataBindingCompilerClasspath().map(_.path),
+      parent = getClass.getClassLoader
+    )
+  }
+
+  def androidDataBindingWorker: Worker[AndroidDataBindingWorker] = Task.Worker {
+    androidDataBindingWorkerClassloader().loadClass(
+      "mill.androidlib.databinding.AndroidDataBindingImpl"
+    )
+      .getConstructor()
+      .newInstance()
+      .asInstanceOf[AndroidDataBindingWorker]
+  }
+
+  def processedLayoutXmls: T[PathRef] = Task {
+
+    val resOutputDir = Task.dest / "resources"
+    val layoutInfoOutputDir = Task.dest / "layout_info"
+
+    os.makeDir.all(resOutputDir)
+    os.makeDir.all(layoutInfoOutputDir)
+    val args = ProcessResourcesArgs(
+      applicationPackageName = androidNamespace,
+      resInputDir = androidResources().head.path.toString,
+      resOutputDir = resOutputDir.toString,
+      layoutInfoOutputDir = layoutInfoOutputDir.toString,
+      enableViewBinding = enableViewBinding,
+      enableDataBinding = enableDataBinding
+    )
+
+    androidDataBindingWorkerModule().processResources(androidDataBindingWorker(), args)
+
+    PathRef(Task.dest)
+  }
+
+  def generatedBindingSources: T[PathRef] = Task {
+    val logDir = Task.dest / "logs"
+    val outputDir = Task.dest / "generated"
+    val classInfoDir = Task.dest / "class_info"
+    os.makeDir.all(logDir)
+    os.makeDir.all(outputDir)
+    os.makeDir.all(classInfoDir)
+    val args = GenerateBindingSourcesArgs(
+      applicationPackageName = androidNamespace,
+      layoutInfoDir = (processedLayoutXmls().path / "layout_info").toString,
+      classInfoDir = classInfoDir.toString,
+      outputDir = outputDir.toString,
+      logFolder = logDir.toString,
+      enableViewBinding = enableViewBinding,
+      enableDataBinding = enableDataBinding
+    )
+
+    androidDataBindingWorkerModule().generateBindingSources(androidDataBindingWorker(), args)
+
+    PathRef(Task.dest)
+  }
+
+  override def generatedSources: T[Seq[PathRef]] = isBindingEnabled match {
+    case true => super.generatedSources() :+ generatedBindingSources()
+    case false => super.generatedSources()
+  }
+
+  override def androidCompiledModuleResources: T[Seq[PathRef]] = isBindingEnabled match {
+    case true => Task {
+        val moduleResources = Seq(processedLayoutXmls().path / "resources")
+
+        val aapt2Compile = Seq(androidSdkModule().aapt2Exe().path.toString(), "compile")
+
+        for (libResDir <- moduleResources) {
+          val segmentsSeq = libResDir.segments.toSeq
+          val libraryName = segmentsSeq.dropRight(1).last
+          val dirDest = Task.dest / libraryName
+          os.makeDir(dirDest)
+          val aapt2Args = Seq(
+            "--dir",
+            libResDir.toString,
+            "-o",
+            dirDest.toString
+          )
+
+          os.call(aapt2Compile ++ aapt2Args)
+        }
+        androidTransitiveCompiledResources() ++ Seq(PathRef(Task.dest))
+      }
+    case false => super.androidCompiledModuleResources()
+  }
 
   override def kotlincPluginMvnDeps: T[Seq[Dep]] = Task {
     val kv = kotlinVersion()
